[#6] refactor: NameSpace 분리
