+++ iOS/TodoList/Util/APIResponse.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+enum APIResponse {
+    static let goodRequest: Int = 200
+    static let itemCreated: Int = 201
+    static let badRequest: Int = 400
+    static let alreadyExist: Int = 409
+}
