+++ iOS/TodoList/Util/NotificationKeyValue.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+enum NotificationKeyValue {
+    static let targetViewTitle = "targetViewTitle"
+    static let targetData = "targetData"
+    static let postTaskData = "postTaskData"
+    static let getTaskData = "getTaskData"
+}
