+++ iOS/TodoList/ViewController/ActionBoardViewController.swift
@@ -10,10 +10,10 @@ class ActionBoardViewController: UIViewController {
     }
     
     private func setupCell() {
-        let nibName = UINib(nibName: NameSpace.nib.actionCardViewCell, bundle: nil)
+        let nibName = UINib(nibName: nibTitle.actionCardViewCell, bundle: nil)
         table.delegate = self
         table.dataSource = self
-        table.register(nibName, forCellReuseIdentifier: NameSpace.identifier.actionCardViewCell)
+        table.register(nibName, forCellReuseIdentifier: identifier.actionCardViewCell)
     }
 
     @IBAction func closeButtonTapped(_ sender: UIButton) {
@@ -27,7 +27,7 @@ extension ActionBoardViewController: UITableViewDelegate, UITableViewDataSource 
     }
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: NameSpace.identifier.actionCardViewCell, for: indexPath) as? ActionCardViewCell else { return UITableViewCell() }
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: identifier.actionCardViewCell, for: indexPath) as? ActionCardViewCell else { return UITableViewCell() }
         return cell
     }
