+++ iOS/TodoList/ViewController/MyTaskViewController.swift
@@ -50,7 +50,7 @@ class MyTaskViewController: UIViewController {
     }
     
     private func editCardButtonTapped(noti: Notification) {
-        guard let popupViewController = storyboard?.instantiateViewController(withIdentifier: NameSpace.identifier.popupViewController) as? EditCardViewController else { return }
+        guard let popupViewController = storyboard?.instantiateViewController(withIdentifier: identifier.popupViewController) as? EditCardViewController else { return }
         popupViewController.modalPresentationStyle = .overFullScreen
         guard let title = noti.userInfo?[NotificationKeyValue.targetViewTitle] as? String else { return }
         popupViewController.targetTitle = title
